Isolate notification candidate selection #158
